// courses-backend/src/common/enums/enums.ts
export enum Role {
  USUARIO = 'USUARIO',
  ALUMNO = 'ALUMNO',
  VENDEDOR = 'VENDEDOR',
  ASISTENTE = 'ASISTENTE',
  PROFESOR = 'PROFESOR',
  PROFESORVENDEDOR = 'PROFESORVENDEDOR',
  ORGANIZADOR = 'ORGANIZADOR',
  FINANZAS = 'FINANZAS',
  ADMIN = 'ADMIN',
}

export enum TipoPago {
  TOTAL = 'TOTAL',
  PARCIAL = 'PARCIAL',
  CUOTA = 'CUOTA',
}

export enum EstadoPedido {
  PENDIENTE = 'PENDIENTE',
  COMPLETADO = 'COMPLETADO',
  CANCELADO = 'CANCELADO',
}

export enum EnrollmentStatus {
  Pendiente = 'Pendiente',
  Parcial = 'Parcial',
  Completo = 'Completo',
  Cancelado = 'Cancelado',
  Inactivo = 'Inactivo',
}